requirements.txt
langgraph
langchain
langsmith
python-dotenv
langchain-groq
langchain-tavily
langgraph-cli[inmem]
langchain_community
faiss-cpu 
pypdf
tiktoken 
langchain-google-genai